refactor: move AgentConfig Neo4j persistence files to agentConfig package
