+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/agentConfig/Neo4jAgentConfigRepository.kt
@@ -1,7 +1,5 @@
-package io.whozoss.agentos.persistence.neo4j
+package io.whozoss.agentos.agentConfig
 
-import io.whozoss.agentos.agentConfig.AgentConfig
-import io.whozoss.agentos.agentConfig.AgentConfigRepository
 import mu.KLogging
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.transaction.annotation.Transactional
@@ -20,7 +18,7 @@ open class Neo4jAgentConfigRepository(
         neo4jRepository
             .save(AgentConfigNode.fromDomain(entity))
             .toDomain()
-            .also { logger.debug { "[Neo4jAgentConfigRepository] Saved agent config ${it.id} ('${entity.name}') under namespace ${entity.namespaceId}" } }
+            .also { logger.debug { "[Neo4jAgentConfigRepository] Saved agent config \${it.id} ('\${entity.name}') under namespace \${entity.namespaceId}" } }
 
     override fun findByIds(ids: Collection<UUID>): List<AgentConfig> =
         neo4jRepository
@@ -47,7 +45,7 @@ open class Neo4jAgentConfigRepository(
     open override fun deleteByParent(parentId: UUID): Int {
         val active = neo4jRepository.findActiveByNamespaceId(parentId.toString())
         neo4jRepository.saveAll(active.map { it.copy(removed = true) })
-        logger.debug { "[Neo4jAgentConfigRepository] Soft-deleted ${active.size} agent configs under namespace $parentId" }
+        logger.debug { "[Neo4jAgentConfigRepository] Soft-deleted \${active.size} agent configs under namespace $parentId" }
         return active.size
     }
