+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/config/Neo4jPersistenceConfiguration.kt
@@ -6,14 +6,14 @@ import io.whozoss.agentos.caseEvent.CaseEventRepository
 import io.whozoss.agentos.caseFlow.CaseRepository
 import io.whozoss.agentos.integrationConfig.IntegrationConfigRepository
 import io.whozoss.agentos.namespace.NamespaceRepository
-import io.whozoss.agentos.persistence.neo4j.AgentConfigNodeNeo4jRepository
+import io.whozoss.agentos.agentConfig.AgentConfigNodeNeo4jRepository
+import io.whozoss.agentos.agentConfig.Neo4jAgentConfigRepository
 import io.whozoss.agentos.persistence.neo4j.CaseEventNodeMapper
 import io.whozoss.agentos.persistence.neo4j.CaseEventNodeNeo4jRepository
 import io.whozoss.agentos.persistence.neo4j.CaseNodeNeo4jRepository
 import io.whozoss.agentos.persistence.neo4j.IntegrationConfigNodeNeo4jRepository
 import io.whozoss.agentos.persistence.neo4j.MessageContentSerializer
 import io.whozoss.agentos.persistence.neo4j.NamespaceNodeNeo4jRepository
-import io.whozoss.agentos.persistence.neo4j.Neo4jAgentConfigRepository
 import io.whozoss.agentos.persistence.neo4j.Neo4jCaseEventRepository
 import io.whozoss.agentos.persistence.neo4j.Neo4jCaseRepository
 import io.whozoss.agentos.persistence.neo4j.Neo4jIntegrationConfigRepository
@@ -49,7 +49,7 @@ import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories
     "'\${agentos.persistence.mode:in-memory}' == 'neo4j' " +
         "or '\${agentos.persistence.mode:in-memory}' == 'embedded-neo4j'",
 )
-@EnableNeo4jRepositories(basePackages = ["io.whozoss.agentos.persistence.neo4j"])
+@EnableNeo4jRepositories(basePackages = ["io.whozoss.agentos.persistence.neo4j", "io.whozoss.agentos.agentConfig"])
 class Neo4jPersistenceConfiguration {
     @Bean
     fun neo4jAgentConfigRepository(agentConfigNodeNeo4jRepository: AgentConfigNodeNeo4jRepository): AgentConfigRepository {
